fix: make lint happy
