+++ pkg/executor/test/analyzetest/BUILD.bazel
@@ -19,7 +19,6 @@ go_test(
         "//pkg/executor",
         "//pkg/infoschema",
         "//pkg/kv",
-        "//pkg/metrics",
         "//pkg/parser/ast",
         "//pkg/parser/mysql",
         "//pkg/parser/terror",
@@ -36,8 +35,6 @@ go_test(
         "//pkg/util/dbterror/exeerrors",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
-        "@com_github_prometheus_client_golang//prometheus",
-        "@com_github_prometheus_client_model//go",
         "@com_github_stretchr_testify//require",
         "@org_uber_go_goleak//:goleak",
     ],
